# docs/api/api.yaml
...
consumes:
- application/json
- application/xml
definitions:
  ApiError:
    properties:
      code:
        description: response code
        format: int64
        type: integer
        x-go-name: Code
      message:
        description: response message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/apisix/manager-api/internal/utils/consts
...
host: 127.0.0.1
info:
  description: Manager API directly operates ETCD and provides data management for APISIX, provides APIs for Front-end or other clients.
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: API doc of Manager API.
paths:
...
  /apisix/admin/consumers:
    get:
      description: Return the consumer list according to the specified page number and page size, and can search by username
      operationId: getConsumerList
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: username of consumer
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: list response
          schema:
            items:
              $ref: '#/definitions/consumer'
            type: array
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
  /apisix/admin/notexist/routes:
    get:
      description: Returns result of route exists checking by name and exclude id
      operationId: checkRouteExist
      parameters:
      - description: name of route
        in: query
        name: name
        type: string
      - description: id of route that exclude checking
        in: query
        name: exclude
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: route not exists
          schema:
            $ref: '#/definitions/ApiError'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
  /apisix/admin/routes:
    get:
      description: Return the route list according to the specified page number and page size, and can search by name and uri
      operationId: getRouteList
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: name of route
        in: query
        name: name
        type: string
      - description: uri of route
        in: query
        name: uri
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: list response
          schema:
            items:
              $ref: '#/definitions/route'
            type: array
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
...
produces:
- application/json
- application/xml
schemes:
- http
- https
swagger: "2.0"